Used `std::cout` instead of `std::printf`.
I had used `std::printf` because when I wrote the example programs, `hdrbg_dump` was actually `memdump`, and private. Hence, I wrote a loop to print a hexadecimal dump of an array of bytes using `std::printf`, which was easier than using `std::cout`, which required additional circus tricks without which it was trying to print them as characters, because `uint8_t` is `char unsigned`. For consistency, I used `std::printf` for the remaining things as well. But since `hdrbg_dump` is now available, I have used `std::cout` now.