Input validation in `hdrbg_uint` and `hdrbg_span`.
This is security-focused library. It is not okay to crash the program or give potentially incorrect results (for `hdrbg_uint` and `hdrbg_span` respectively). I introduced more errors to signal such issues. Updated the tests to verify everything (and enhanced existing tests as well). Made the necessary changes to the Python API. Anothing point is that the type of `r_length` in `hdrbg_fill` is now `int long unsigned`, because the previous type, `size_t`, may not be large enough to store 65536 (`SIZE_MAX` can be 65535).