`hdrbg_fill` should always advance the state.
According to the specification, the state is advanced even if a request is made for zero bytes. This implementation wasn't advancing the state for such requests. I have changed it to match the specification. Also, I used this opportunity to implement `hdrbg_drop`, a function which does the same thing as similarly-named functions in my Mersenne twister implementation.